Use the same base test for serializers as they all should return the exact same format
