+++ tests/Bernard/Tests/Serializer/JMSSerializerTest.php
@@ -2,70 +2,17 @@
 
 namespace Bernard\Tests\Serializer;
 
-use Bernard\Message\Envelope;
-use Bernard\Message\DefaultMessage;
 use Bernard\Serializer\JMSSerializer;
+use JMS\Serializer\SerializerBuilder;
 
-class JMSSerializerTest extends \PHPUnit_Framework_TestCase
+class JMSSerializerTest extends AbstractSerializerTest
 {
-    public function setUp()
-    {
-        $this->serializer = new JMSSerializer($this->createJMSSerializer());
-    }
-
-    public function testImplementsSerializer()
-    {
-        $this->assertInstanceOf('Bernard\Serializer', $this->serializer);
-    }
-
-    public function testItSerializesDefaultMessage()
-    {
-        $json = '{"args":{"name":"SendNewsletter"},"class":"Bernard:Message:DefaultMessage","timestamp":' . time() . ',"retries":0}';
-        $this->assertEquals($json, $this->serializer->serialize($this->createWrappedDefaultMessage('SendNewsletter')));
-
-        $json = '{"args":{"name":"SendNewsletter","newsletterId":1,"users":["henrikbjorn"]},"class":"Bernard:Message:DefaultMessage","timestamp":' . time() . ',"retries":0}';
-        $this->assertEquals($json, $this->serializer->serialize($this->createWrappedDefaultMessage('SendNewsletter', array(
-            'newsletterId' => 1,
-            'users' => array(
-                'henrikbjorn'
-            ),
-        ))));
-    }
-
-    public function testItSerializesACustomImplementedMessage()
-    {
-        $json = '{"args":{},"class":"Application:SendNewsletterMessage","timestamp":' . time() . ',"retries":0}';
-        $this->assertEquals($json, $this->serializer->serialize(new Envelope(new \Application\SendNewsletterMessage())));
-    }
-
-    public function testItDeserializesACustomImplementedMessage()
-    {
-        $json = '{"args":{},"class":"Application:SendNewsletterMessage","timestamp":' . time() . ',"retries":0}';
-        $this->assertEquals($json, $this->serializer->serialize(new Envelope(new \Application\SendNewsletterMessage())));
-    }
-
-    public function testItDeserializesDefaultMessage()
-    {
-        $message = $this->createWrappedDefaultMessage('SendNewsletter');
-        $json = $this->serializer->serialize($message);
-
-        $this->assertEquals($message, $this->serializer->deserialize($json));
-    }
-
-    public function createWrappedDefaultMessage($name, array $properties = array())
-    {
-        return new Envelope(new DefaultMessage($name, $properties));
-    }
-
-    public function createJMSSerializer()
+    public function createSerializer()
     {
         $class = new \ReflectionClass('Bernard\Serializer');
-        $builder = new \JMS\Serializer\SerializerBuilder();
+        $builder = new SerializerBuilder();
         $builder->addMetadataDir(dirname($class->getFilename()) . '/Resources/serializer', 'Bernard');
 
-        return $builder->build();
+        return new JMSSerializer($builder->build());
     }
 }
-
-namespace Application;
-class SendNewsletterMessage extends \Bernard\Message\AbstractMessage {}
